+++ justine/rcemu/src/traffic.cpp
@@ -205,8 +205,8 @@ void justine::robocar::Traffic::cmd_session ( boost::asio::ip::tcp::socket clien
                   std::shared_ptr<SmartCar> c = m_smart_cars_map[cl.get_id()];
 
                   length += std::sprintf ( data+length,
-                                           "<OK %d %u %u %u>", cl.get_id(), c->from(),
-                                           c->to_node(), c->get_step() );
+                                           "<OK %d %u %u %u>", cl.get_id(), (unsigned int)c->from(),
+                                           (unsigned int)c->to_node(), (unsigned int)c->get_step() );
                 }
               else
                 length += std::sprintf ( data+length, "<ERR unknown car id>" );
@@ -227,8 +227,8 @@ void justine::robocar::Traffic::cmd_session ( boost::asio::ip::tcp::socket clien
                       if ( c->get_type() == CarType::GANGSTER )
                         {
                           length += std::sprintf ( data+length,
-                                                   "<OK %d %u %u %u>", cl.get_id(), c->from(),
-                                                   c->to_node(), c->get_step() );
+                                                   "<OK %d %u %u %u>", cl.get_id(), (unsigned int)c->from(),
+                                                   (unsigned int)c->to_node(), (unsigned int)c->get_step() );
 
                           if ( length > network_buffer_size - 512 )
                             {
@@ -260,8 +260,8 @@ void justine::robocar::Traffic::cmd_session ( boost::asio::ip::tcp::socket clien
                   for ( auto c:m_cop_cars )
                     {
                       length += std::sprintf ( data+length,
-                                               "<OK %d %u %u %u %d>", cl.get_id(), c->from(),
-                                               c->to_node(), c->get_step(), c->get_num_captured_gangsters() );
+                                               "<OK %d %u %u %u %d>", cl.get_id(), (unsigned int)c->from(),
+                                               (unsigned int)c->to_node(), (unsigned int)c->get_step(), c->get_num_captured_gangsters() );
 
                       if ( length > network_buffer_size - 512 )
                         {
